this commit adds the call to `luksErase`, which will (with --batch-mode) make all of the encrypted data on that luks container permanently unrecoverable, unless a backup of the header with the keyslots (or the decrypted master key) were restored.
If that `luksErase` fails for whatever reason (say a machine's
`cryptsetup` is older than 2014), then it will attempt to just overwrite
the first 20 MB of the device with disk destroyer.

My tests show, however, that the `luksErase` does *not* overwrite the plaintext
metadata on the luks container, so I'm adding a TODO item to wipe that
out too.