I finished the plaintext metadata overwrite in the LUKS header after luksErase is called. This has been tested on a luks container in a file only; I haven't actually tested this on a block device wipe yet.
For testing, I've been using `bgrep` to search the LUKS container for a
sequence of bytes (in hex) that are part of the LUKS standard. This
works better for LUKS2 since it's variable and the metadata is stored in
metadata with a set of mandatory JSON attributes that *must* be written
to the LUKS header--wheras LUKS1 is just implicitly encoded and
therefore not as easy to distinguish from random data.

 * https://github.com/tmbinc/bgrep

Here's the space-seperated list of byte sequences that I'm using so far
(in hex).

sequences="4c554b53babe 4c554b53 6c756b73 6b6579736c6f74 737472697065 6f6666736574 68617368 61726561 656e6372797074 73616c74 7365676d656e74 6a736f6e"

That's the LUKS 0xbabe magic header, LUKS, luks, keyslot, stripe,
offset, hash, area, encrypt, salt, segment, and json.

Note that I had a larger set (including aes, sha, "}, {", etc) but I
found too many false-positives with 2-byte and 3-byte lengths.

Here's the quick one-liner command I'm using with `bgrep` to validate
the shredding of the plaintext metadata with this script so far:

device=luksVol2; for s in ${sequences}; do offset=`bgrep ${s} ${device} | cut -d' ' -f2`; echo -e "${s}\n----------"; for o in ${offset}; do hexdump -Cs "0x${o}" -n16 ${device}; done; echo '=========='; done

Next I'd like to actually test this self-destruct sequence on a
disposable OS, but I'm afraid that cryptsetup, head, and all the other
commands will be inaccessible as soon as the `luksSuspned` is called. So
the next logical TODO is actually to setup the ramfs just before the
`luksSuspend` with all the tools needed to finish the execution of this
script.